requirements.txt
fastapi
uvicorn
chromadb
unstructured
sentence-transformers
python-dotenv
